New Updates

+++ Hafta1-MertKarakaya/Models/TutarModel.cs
@@ -7,7 +7,7 @@ namespace Hafta1_MertKarakaya.Models
 {
     public class Tutar
     {
-        public long geriOdenecektoplamTutar { get; set; }
-        public long odenecekFaizTutari { get; set; }
+        public double geriOdenecektoplamTutar { get; set; }
+        public double odenecekFaizTutari { get; set; }
     }
 }
